Integration Auth. Module. - working version
